Updated by Kamalakar Dandu

+++ tdp_resolution/src/main/java/com/itgrids/dao/ITdpResolutionMailDAO.java
@@ -4,7 +4,7 @@ import org.appfuse.dao.GenericDao;
 
 import com.itgrids.model.TdpResolutionMail;
 
-public interface ITdpResolutionMailDAO extends GenericDao<TdpResolutionMail,Integer> {
+public interface ITdpResolutionMailDAO extends GenericDao<TdpResolutionMail,Long> {
